Move deepFreeze to a testUtils file
